chore(test): audit and clean up test suite and QA scripts
Merge synonymous test files (ToolWindow streaming parameterized,
StatusMonitorModal ACP status, floating-window/session-tree/update-transport
single-case suites), delete orphaned QA scripts (scroll-anchor,
settings-layout + fixtures) and static source-string-only tests, strengthen
weak assertions (locale key parity/non-empty, configOptions error handling,
forgeCommands completeness), and fix misleading describe names across
utils/composables/backends suites. Refresh stale historical references in QA
script comments and harden the Windows bridge installer QA script with
deterministic artifact selection and try/finally cleanup.
