+++ app/composables/useBackendSessionTrees.archiveCache.test.ts
@@ -3,7 +3,7 @@ import { nextTick, reactive, ref } from 'vue';
 import type { ProjectState } from '../types/worker-state';
 import { useBackendSessionTrees } from './useBackendSessionTrees';
 
-describe('useBackendSessionTrees archive cache invalidation', () => {
+describe('useBackendSessionTrees projections and archive cache', () => {
   it('rebuilds TopPanel data immediately when a session archive value changes', async () => {
     const projects = reactive<Record<string, ProjectState>>({
       acp: {
@@ -314,5 +314,46 @@ describe('useBackendSessionTrees archive cache invalidation', () => {
         ],
       }),
     ]);
+  });
+
+  it('projects an archived OpenCode session into the TopPanel archive state', () => {
+    // Given
+    const projects = reactive<Record<string, ProjectState>>({
+      opencode: {
+        id: 'opencode',
+        name: 'OpenCode',
+        worktree: '/repo',
+        sandboxes: {
+          '/repo': {
+            directory: '/repo',
+            name: 'repo',
+            rootSessions: ['session-1'],
+            sessions: {
+              'session-1': {
+                id: 'session-1',
+                title: 'Archived session',
+                directory: '/repo',
+                timeUpdated: 1,
+                timeArchived: 123,
+              },
+            },
+          },
+        },
+      },
+    });
+
+    // When
+    const trees = useBackendSessionTrees({
+      activeBackendKind: ref('opencode'),
+      projects,
+      pinnedStore: ref({}),
+      deletedSandboxStore: ref({}),
+      homePath: ref('/home/test'),
+      replaceHomePrefix: (path) => path,
+      resolveProjectColor: () => undefined,
+    });
+
+    // Then
+    expect(trees.topPanelTreeData.value[0]?.sandboxes[0]?.sessions[0]?.archivedAt).toBe(123);
   });
 });
