--- app/dev/settings-layout-fixture.html
@@ -1,24 +1,0 @@
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta name="viewport" content="width=device-width,initial-scale=1" />
-    <meta
-      http-equiv="Content-Security-Policy"
-        content="default-src 'self'; script-src 'self' 'unsafe-inline'; style-src 'self' 'unsafe-inline'; connect-src 'self' http://localhost:* http://127.0.0.1:* ws://localhost:* ws://127.0.0.1:* https://api.iconify.design; img-src 'self' data: blob:; font-src 'self' data:; worker-src 'self' blob:;"
-    />
-    <title>Settings Layout Fixture (dev-only, mocked IPC)</title>
-    <script type="module" src="./settings-layout-fixture.ts"></script>
-    <style>
-      html,
-      body {
-        margin: 0;
-        padding: 0;
-        background: #0d1117;
-      }
-    </style>
-  </head>
-  <body>
-    <div id="app"></div>
-  </body>
-</html>
